bugfix: models.is_complete() and models.is_depictable() throw an exception
Both models.is_complete() and models.is_depictable()  throw an exception claiming the class Problem is not an iterable. The reason seems to be the use of the `in` operator, as the appropriate „magic method“ is not implemented in the class and the standard operator expects an iterable. Basically, the methods test if some attributes are set, so I just implemented that as testing the attributes for being „None / not None“.